feat: multiple improvements
