--- .github/workflows/main.yml
@@ -1,99 +1,0 @@
-name: Release Creation
-
-on:
-  release:
-    types: [published]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-
-    # Extract tag version number
-    - name: Extract tag version number
-      id: get_version
-      uses: battila7/get-version-action@v2
-
-    # Set up Node.js
-    - uses: actions/setup-node@v3
-      with:
-        node-version: '18'
-    - run: npm install
-
-    # Build Release
-    - name: Generate release
-      run: mkdir styles && mkdir css && npm run release
-
-    # Substitute Manifest and Download Links For Versioned Ones
-    - name: Substitute Manifest and Download Links For Versioned Ones
-      id: sub_manifest_link_version
-      uses: microsoft/variable-substitution@v1
-      with:
-        files: ./release/system.json
-      env:
-        version: ${{steps.get_version.outputs.version-without-v}}
-        url: https://github.com/${{github.repository}}
-        manifest: https://github.com/${{github.repository}}/releases/latest/download/system.json
-        download: https://github.com/${{github.repository}}/releases/download/${{github.event.release.tag_name}}/system.zip
-
-    # Create a zip file with all files required by the system to add to the release
-    - run: cd release && zip -r ./system.zip system.json template.json readme.md LICENSE.txt assets/ build/ fonts/ lang/ templates/ packs/ readme/ css/
-
-    # Update Release with Files
-    - name: Update Release with Files
-      id: create_version_release
-      uses: ncipollo/release-action@v1
-      with:
-        allowUpdates: true # Set this to false if you want to prevent updating existing releases
-        name: ${{ github.event.release.name }}
-        draft: ${{ github.event.release.unpublished }}
-        prerelease: ${{ github.event.release.prerelease }}
-        token: ${{ secrets.REPO_SEC }}
-        artifacts: './release/system.json, ./release/system.zip'
-        tag: ${{ github.event.release.tag_name }}
-        body: ${{ github.event.release.body }}
-    # Publish version to foundry
-    - name: Get System ID
-      id: systemID
-      uses: notiz-dev/github-action-json-property@release
-      with:
-        path: "./release/system.json"
-        prop_path: "id"
-
-    - name: Get minimum
-      id: minimum
-      uses: notiz-dev/github-action-json-property@release
-      with:
-        path: "./release/system.json"
-        prop_path: "compatibility.minimum"
-
-    - name: Get verified
-      id: verified
-      uses: notiz-dev/github-action-json-property@release
-      with:
-        path: "./release/system.json"
-        prop_path: "compatibility.verified"
-  
-    # Submit to foundry
-    - name: Submit package to FoundryVTT Package Release API
-      run: |
-        curl -X POST "https://foundryvtt.com/_api/packages/release_version/" \
-          -H "Content-Type: application/json" \
-          -H "Authorization: ${{ secrets.FOUNDRY_KEY }}" \
-          -d '{
-            "id": "${{ steps.systemID.outputs.prop }}",
-            "dry-run": false,
-            "release": {
-              "version": "${{ steps.get_version.outputs.version-without-v }}",
-              "manifest": "https://github.com/${{ github.repository }}/releases/latest/download/system.json",
-              "notes": "https://github.com/${{ github.repository }}/releases/tag/${{ steps.get_version.outputs.version-without-v }}",
-              "compatibility": {
-                "minimum": "${{ steps.minimum.outputs.prop }}",
-                "verified": "${{ steps.verified.outputs.prop }}",
-                "maximum": ""
-              }
-            }
-          }'
-
-
