+++ module/applications/sheets/character-sheet.mjs
@@ -1,15 +1,9 @@
-import {
-  onManageActiveEffect,
-  prepareActiveEffectCategories,
-} from '../helpers/effects.mjs';
-import { HOUSEHOLD } from '../helpers/config.mjs';
-import { addProfession } from "../helpers/professions.mjs";
-import * as actions from "../helpers/actions.mjs";
+import { HOUSEHOLD } from '../../helpers/config.mjs';
+import { addProfession } from "../../helpers/professions.mjs";
+import * as actions from "../../helpers/actions.mjs";
+import { HouseholdBaseActorSheet } from "./base-actor-sheet.mjs";
 
-const { HandlebarsApplicationMixin } = foundry.applications.api;
-const { ActorSheetV2 } = foundry.applications.sheets
-
-export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2) {
+export class HouseholdActorSheet extends HouseholdBaseActorSheet {
 
   /* -------------------------------------------- */
   /*  Configuration                               */
@@ -34,7 +28,7 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
       actionConfig: this.prototype._onActionConfig,
       customEdit: this.prototype._onCustomEdit,
       effectControl: this.prototype._onEffectControl,
-      editImage: this.#onEditImage
+      editImage: this._onEditImage
     },
     form: {
       submitOnChange: true,
@@ -98,56 +92,14 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
   /* -------------------------------------------- */
 
   async _prepareContext(options) {
-    const context = await super._prepareContext(options);
-    
-    // Use this.actor for ActorSheetV2
-    if (this.token) {
-      context.actor = this.token.actor;
-      context.id = this.token.id;
-    } else {
-      context.actor = this.actor;
-      context.id = this.actor.id;
-    }
-    
-    context.system = context.actor.system;
-    context.flags = context.actor.flags;
-    context.items = context.actor.items.contents;
-    context.effects = prepareActiveEffectCategories(
-      context.actor.allApplicableEffects()
-    );
-    context.editable = this.isEditable;
-    context.rollData = context.actor.getRollData();
-
-    // Prepare tabs configuration
-    context.tabs = this._prepareTabs("primary");
-    
-
-    if (context.actor.type === 'character') {
-      this._prepareItems(context);
-      this._prepareCharacterData(context);
-    }
-
-    if (context.actor.type === 'npc' || context.actor.type === 'opponent') {
-      this._prepareItems(context);
-    }
-
-    const enrichmentOptions = {
-      secrets: this.document.isOwner,
-      relativeTo: this.document
-    };
-    // Enrich description for all item types
-    if (this.document.system.biography) {
-      context.descriptionHTML = await foundry.applications.ux.TextEditor.implementation.enrichHTML(
-        this.document.system.biography,
-        enrichmentOptions
-      );
-    }
-
+    const context = await this._prepareBaseContext(options);
+    this._prepareItems(context);
+    this._prepareCharacterData(context);
     return context;
   }
 
   /* -------------------------------------------- */
-  /*  Item & Character Prep                       */
+  /*  Character Prep                              */
   /* -------------------------------------------- */
 
   _prepareCharacterData(context) {
@@ -172,79 +124,9 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
     }
   }
 
-  _prepareItems(context) {
-    const gear = [];
-    const weapons = [];
-    const moves = [];
-    const gadgets = [];
-    const contracts = [];
-    const traits = [];
-
-    for (const i of context.items) {
-      i.img ||= Item.DEFAULT_ICON;
-
-      if (i.type === 'item') gear.push(i);
-      else if (i.type === 'weapon') weapons.push(i);
-      else if (i.type === 'gadget') gadgets.push(i);
-      else if (i.type === 'move') moves.push(i);
-      else if (i.type === 'contract') contracts.push(i);
-      else if (i.type === 'trait') traits.push(i);
-    }
-
-    context.gear = gear;
-    context.weapons = weapons;
-    context.traits = traits;
-    context.gadgets = gadgets;
-    context.moves = moves;
-    context.contracts = contracts;
-    context.features = [];
-    context.all_items = [...traits, ...moves, ...contracts, ...gadgets, ...weapons, ...gear];
-  }
-
   /* -------------------------------------------- */
-  /*  Rendering                                   */
+  /*  Actions                                     */
   /* -------------------------------------------- */
-
-  /**
-   * Attach event listeners after rendering
-   * @param {ApplicationRenderContext} context  Render context
-   * @param {RenderOptions} options            Render options
-   * @override
-   */
-  _onRender(context, options) {
-    super._onRender(context, options);
-
-    // Attach right-click listener to profile image
-    const profileImg = this.element.querySelector('.profile-img');
-    if (profileImg) {
-      profileImg.addEventListener('contextmenu', this.constructor.#onImageRightClick.bind(this));
-    }
-  }
-
-  /* -------------------------------------------- */
-  /*  Actions (replaces activateListeners)        */
-  /* -------------------------------------------- */
-
-  _onItemEdit(event, target) {
-    
-    const item = this.document.items.get(target.dataset.itemId);
-    item?.sheet?.render(true);
-  }
-
-  async _onItemDelete(event, target) {
-    const item = this.document.items.get(target.dataset.itemId);
-    await item?.delete();
-  }
-
-  async _onItemCreate(event, target) {
-    const type = target.dataset.type;
-    const name = `New ${type.capitalize()}`;
-    await Item.create({
-      name,
-      type,
-      system: {}
-    }, { parent: this.document });
-  }
 
   async _onActionConfig() {
     const actor = this.document;
@@ -286,8 +168,8 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
     const actor = this.document;
     if (target.dataset.object === 'actor') {
       const { path, value, dtype, object } = target.dataset;
-      
-      
+
+
 
       // Special handling for conditions - call toggleCondition
       if (path.includes('system.conditions.')) {
@@ -304,26 +186,16 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
         newValue = Number(value);
       }
 
-      
-      
+
+
       await actor.update({ [path]: newValue });
     } else if (target.dataset.object === 'item') {
       let ev = {}
       ev.currentTarget = target;
-      
+
       ev.currentTarget.dataset.action = "use";
       await actions.useItem(ev);
     }
-  }
-
-  _onEffectControl(event, target) {
-    const li = target.closest("li");
-    const parent =
-      li.dataset.parentId === this.document.id
-        ? this.document
-        : this.document.items.get(li.dataset.parentId);
-
-    onManageActiveEffect(event, parent);
   }
 
   /* -------------------------------------------- */
@@ -333,7 +205,7 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
   async _onRoll(event, target) {
     const actor = this.document;
     const dataset = target.dataset;
-    
+
 
     if (dataset.type === 'action') {
       const roll = await new Roll("1d6", actor.getRollData()).evaluate();
@@ -343,24 +215,24 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
         rollMode: game.settings.get('core', 'rollMode')
       });
     } else if (dataset.type === 'skill') {
-      
+
       this.actor.dialogRollSkill(dataset);
 
     } else if (dataset.type == 'attack') {
-      
-      
+
+
       const item_id = target.closest('.item-list')?.dataset.itemId;
-      
+
       const item = this.actor.items.get(item_id);
       if (item) {
-        
+
 
         dataset.label = item.system.field;
         dataset.field = item.system.field;
         dataset.key = item.system.skill;
         dataset.itemId = item.id;
         dataset.characterId = this.actor.id;
-        
+
         this.actor.dialogRollSkill(dataset);
 
 
@@ -376,69 +248,6 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
         });
       }
     }
-  }
-
-  _onShow(event, target) {
-    //this.document.dialogRollSkill(target.dataset);
-    
-    let forward_event = {};
-    forward_event.currentTarget = target;
-    if (target.dataset?.subAction) {
-      const sub_action = target.dataset.subAction;
-      const parent = target.closest('.item-list')?.dataset;
-      
-      if (parent) {
-        forward_event.currentTarget.dataset.action = sub_action;
-        forward_event.currentTarget.dataset.itemId = parent.itemId;
-        forward_event.currentTarget.dataset.characterId = parent.characterId;
-        forward_event.currentTarget.dataset.object = parent.object;
-      }
-
-    }
-    
-    actions.useItem(forward_event);
-  }
-
-  /**
-   * Handle click event to edit the actor image
-   * @param {PointerEvent} event  The originating click event
-   * @param {HTMLElement} target  The clicked element
-   * @private
-   */
-  static async #onEditImage(event, target) {
-    const field = target.dataset.field || "img";
-    const current = foundry.utils.getProperty(this.document, field);
-
-    const fp = new foundry.applications.apps.FilePicker({
-      type: "image",
-      current: current,
-      callback: (path) => this.document.update({ [field]: path })
-    });
-
-    fp.render(true);
-  }
-
-  /**
-   * Handle right-click on actor image to view or share
-   * @param {MouseEvent} event  The originating contextmenu event
-   * @private
-   */
-  static #onImageRightClick(event) {
-    event.preventDefault();
-
-    const ip = new foundry.applications.apps.ImagePopout({
-      src: this.actor.img,
-      uuid: this.actor.uuid,
-      window: { title: this.actor.name }
-    });
-
-    // Display the image popout
-    ip.render(true);
-
-    // If GM, also share the image with other players
-    // if (game.user.isGM) {
-    //   ip.shareImage();
-    // }
   }
 
   /* -------------------------------------------- */
@@ -494,7 +303,7 @@ export class HouseholdActorSheet extends HandlebarsApplicationMixin(ActorSheetV2
     }
 
     if (data.type !== "Item") return;
-    
+
     // await this.document.createEmbeddedDocuments("Item", [{
     //   name: item.name,
     //   type: item.type,
