+++ module/documents/_module.mjs
@@ -1,0 +1,4 @@
+export { HouseholdActor } from "./actor.mjs";
+export { HouseholdItem } from "./item.mjs";
+export { HouseholdCombat } from "./combat.mjs";
+export { HouseholdCombatant } from "./combatant.mjs";
