+++ module/helpers/macros.mjs
@@ -1,0 +1,65 @@
+/* -------------------------------------------- */
+/*  Hotbar Macros                               */
+/* -------------------------------------------- */
+
+/**
+ * Create a Macro from an Item drop.
+ * Get an existing item macro if one exists, otherwise create a new one.
+ * @param {Object} data     The dropped data
+ * @param {number} slot     The hotbar slot to use
+ * @returns {Promise}
+ */
+export async function createItemMacro(data, slot) {
+  // First, determine if this is a valid owned item.
+  if (data.type !== 'Item') return;
+  if (!data.uuid.includes('Actor.') && !data.uuid.includes('Token.')) {
+    return ui.notifications.warn(
+      'You can only create macro buttons for owned Items'
+    );
+  }
+  // If it is, retrieve it based on the uuid.
+  const item = await Item.fromDropData(data);
+
+  // Create the macro command using the uuid.
+  const command = `game.household.rollItemMacro("${data.uuid}");`;
+  let macro = game.macros.find(
+    (m) => m.name === item.name && m.command === command
+  );
+  if (!macro) {
+    macro = await Macro.create({
+      name: item.name,
+      type: 'script',
+      img: item.img,
+      command: command,
+      flags: { 'household.itemMacro': true },
+    });
+  }
+  game.user.assignHotbarMacro(macro, slot);
+  return false;
+}
+
+/**
+ * Create a Macro from an Item drop.
+ * Get an existing item macro if one exists, otherwise create a new one.
+ * @param {string} itemUuid
+ */
+export function rollItemMacro(itemUuid) {
+  // Reconstruct the drop data so that we can load the item.
+  const dropData = {
+    type: 'Item',
+    uuid: itemUuid,
+  };
+  // Load the item from the uuid.
+  Item.fromDropData(dropData).then((item) => {
+    // Determine if the item loaded and if it's an owned item.
+    if (!item || !item.parent) {
+      const itemName = item?.name ?? itemUuid;
+      return ui.notifications.warn(
+        `Could not find item ${itemName}. You may need to delete and recreate this macro.`
+      );
+    }
+
+    // Trigger the item roll
+    item.roll();
+  });
+}
